fixes and cpp workshop

+++ backend/UniTodoSpring/src/main/java/com/teamviewer/uni_todo/UniTodoSpring/repositories/TodoRepository.java
@@ -3,5 +3,5 @@ package com.teamviewer.uni_todo.UniTodoSpring.repositories;
 import com.teamviewer.uni_todo.UniTodoSpring.domains.TodoDomain;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TodoRepository extends CrudRepository<TodoDomain, Long> {
+public interface TodoRepository extends CrudRepository<TodoDomain, Integer> {
 }
